requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
python-dotenv==0.19.2
aiohttp==3.8.1
sqlalchemy==1.4.39
PyJWT==1.7.1
email-validator==1.2.1